Add win/lose screen and transition to main menu.

+++ Assets/Scripts/Base.cs
@@ -7,14 +7,14 @@ public class Base : MonoBehaviour
 {
     [SerializeField] private bool isEnemyBase = false;
     
-    public static event Action<float> OnEnemyEnterBase;
+    public static event Action<int> OnEnemyEnterBase;
     public static event Action OnWinGame;
     
     private void OnTriggerEnter2D(Collider2D col)
     {
         if (col.gameObject.CompareTag("Enemy"))
         {
-            OnEnemyEnterBase?.Invoke(10f);
+            OnEnemyEnterBase?.Invoke(10);
             Destroy(col.gameObject);
         }
